move all the template related code to a seperate file
